// DigitalHealth/src/app/models/healt-model.ts
export interface Health {
  referenceId?: String;
  schmerzenRuecken: number;

  schmerzenNacken: number;

  schmerzenHandgelenk: number;

  schmerzenKnie: number;

  schmerzenHuefte: number;

  schmerzenSchulter: number;

  schmerzenFussgelenke: number;

  toggelBeweglichket: Boolean;

  schmerzenBrennen: Boolean;

  schmerzenKribbeln: Boolean;

  schmerzenTaubheit: Boolean;

  schmerzenUeberempfindlichkeit: Boolean;

  toggelkraft: Boolean;

  schmerzenNadeln: Boolean;

  schmerzenFeld: String;

  schmerzenArbeiten: Boolean;

  schmerzenLaufen: Boolean;

  schmerzenBuecken: Boolean;

  schmerzenStress: Boolean;

  schmerzenWetter: Boolean;

  schmerzenSpringen: Boolean;
}